Revert "[Turbopack] Move DirList to its own module (#88256)"
This reverts commit d90847dead1f5c135741043f7720844583a26d0a.
